requirements.txt
# Core AI and ML
torch>=2.0.0
transformers>=4.30.0
vllm==0.2.7
accelerate>=0.20.0
bitsandbytes>=0.41.0

# LiveKit Integration
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.0
livekit-api>=0.5.0
livekit-protocol>=0.5.0

# Voice Processing
deepgram-sdk>=3.0.0
elevenlabs>=0.2.0
openai>=1.0.0
webrtcvad>=2.0.0
speechrecognition>=3.10.0
pydub>=0.25.0
librosa>=0.10.0

# Real-time Communication
websockets>=11.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
socketio>=5.8.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
httpx>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0
python-json-logger>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Utilities
requests>=2.31.0
aiofiles>=23.1.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# GPU Monitoring
gpustat>=1.1.0
nvidia-ml-py>=12.535.0

# Security
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Audio Processing
soundfile>=0.12.0
pyaudio>=0.2.11
wave>=0.0.2
